scope the role down to only allow delete specific log group

+++ packages/@aws-cdk/aws-logs/lib/log-retention-provider/index.ts
@@ -156,12 +156,9 @@ export async function handler(event: AWSLambda.CloudFormationCustomResourceEvent
       }
     }
 
-    //if the requestType is delete, then delete the log group created by the logRetention
-    if (event.RequestType === 'Delete') {
-      //if the removal policy is delete, then delete the log group
-      if (logDeletionPolicy === 'destroy') {
-        await deleteLogGroup(logGroupName, logGroupRegion, retryOptions);
-      }
+    //When the requestType is delete, delete the log group if the removal policy is delete
+    if (event.RequestType === 'Delete' && logDeletionPolicy === 'destroy') {
+      await deleteLogGroup(logGroupName, logGroupRegion, retryOptions);
       //else retain the log group
     }
